fixed some warnings
w1: int sum used in a floating point context, w2: changed %d to %f for per

+++ marks.c
@@ -6,9 +6,9 @@ int main() {
   printf("Enter Marks for 5 subjects:\n");
   scanf("%d%d%d%d%d", &s1, &s2, &s3, &s4, &s5);
   sum = s1 + s2+ s3+ s4+ s5;
-  per = sum/5;
+  per = (float)sum/5;
 
   printf("sum is: %d\n", sum);
-  printf("percentage is: %d\n", per);
+  printf("percentage is: %f\n", per);
   return 0;
 }
